% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calcula.R
\name{calcula}
\alias{calcula}
\title{calcula}
\usage{
calcula()
}
\value{
Retorna uma lista com as tabelas úteis para análises das emissões de GEE nos
planos de energia
}
\description{
Calcula as emissões de gases de efeito estufa a partir da matriz
consolidada produzida pela SEE em formato excel. A função primeiramente organiza
as informações da planilha em um dataframe e depois procede os cálculos, gerando
tabelas úteis para análises dos resultados.
O arquivo excel é importado após o usuário clicar no botão calcular - código
de importação no script app_server.R.
Os fatores de emissão são importados no script global.R
}
