% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exportar.R
\name{exportar}
\alias{exportar}
\title{exportar}
\usage{
exportar()
}
\value{
Arquivo excel contendo as tabelas úteis organizadas
}
\description{
Exporta as tabelas úteis geradas pela execução da função 'calcula' para excel
}
